requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
streamlit==1.44.1
requests==2.32.3
python-dotenv==1.0.1
python-multipart==0.0.20

llama-index==0.12.28
llama-index-vector-stores-qdrant==0.4.3
llama-index-embeddings-huggingface==0.5.3
llama-index-llms-openai-like==0.3.4
llama-index-llms-gemini==0.4.8
qdrant-client>=1.14.0
sentence-transformers==3.4.1

pypdf==5.4.0
python-docx==1.1.2
